producer/requirements.txt
astroid==2.3.3
Click==7.0
DateTime==4.3
Flask==1.1.1
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
kafka-python==1.4.7
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.17.4
opencv-contrib-python==4.1.2.30
pylint==2.4.4
pytz==2019.3
six==1.13.0
typed-ast==1.4.0
u-msgpack-python==2.5.2
Werkzeug==0.16.0
wrapt==1.11.2
zope.interface==4.7.1
